requirements.txt
# Зависимости
fastapi==0.128.0
uvicorn[standard]==0.40.0
sqlalchemy[asyncio]==2.0.45
python-dotenv==1.2.1
pydantic==2.12.5
anyio==4.12.1
starlette==0.50.0
typing_extensions==4.15.0
asyncpg==0.29.0